Add warnings for missing RC/IASL
My network's shit so it really has a tendency to run without it. The download error display is somehow never useful, so add warnings at compile time.

We could be more helpful and say stuff like "do this to compile without going through the selection", but lazy lazy stuff boring.